+++ packages/gu-ds-components/src/progress-bar/react.ts
@@ -1,3 +1,3 @@
-import { ProgressBar } from ".";
+import { ProgressBar } from "./ProgressBar";
 
 export default ProgressBar;
