+++ packages/gu-ds-components/src/progress-bar/web-component.ts
@@ -1,5 +1,5 @@
 import r2wc from "@r2wc/react-to-web-component";
-import { ProgressBar, Props, propNames } from ".";
+import { ProgressBar, Props, propNames } from "./ProgressBar";
 
 const component = r2wc<Props>(ProgressBar, {
   props: propNames,
